Move satpy/composites/__init__.py tests to tests/test_composites.py
