+++ .codebeatignore
@@ -1,2 +1,1 @@
 satpy/version.py
-versioneer.py
